Move StackingContextHelpers from Frame -> Scene
